Score and enemy random generation added
Few of the bugs have been removed and the enemy fall is completely
implemented to spawn at random positions. And as soon as a enemy strikes
the floor 1 point will be added to the score. Some unnecessary files
have been deleted. Do try this and give us the feedback about any bugs.